requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
dj-database-url==2.0.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
